refactor: 클래스명 Strategy에서 Policy로 변경
