+++ src/main/java/janggi/domain/piece/BasePiece.java
@@ -1,14 +1,14 @@
 package janggi.domain.piece;
 
 import janggi.domain.Side;
-import janggi.domain.strategy.MoveStrategy;
+import janggi.domain.policy.RoutePolicy;
 
 public abstract class BasePiece implements Piece {
-    protected MoveStrategy moveStrategy;
+    protected RoutePolicy routePolicy;
     protected Side side;
 
-    public BasePiece(MoveStrategy moveStrategy, Side side) {
-        this.moveStrategy = moveStrategy;
+    public BasePiece(RoutePolicy routePolicy, Side side) {
+        this.routePolicy = routePolicy;
         this.side = side;
     }
