+++ src/main/java/janggi/domain/piece/Cha.java
@@ -1,10 +1,10 @@
 package janggi.domain.piece;
 
 import janggi.domain.Side;
-import janggi.domain.strategy.DefaultMoveStrategy;
+import janggi.domain.policy.ClearPathPolicy;
 
 public class Cha extends LinearPiece {
     public Cha(Side side) {
-        super(new DefaultMoveStrategy(), side);
+        super(new ClearPathPolicy(), side);
     }
 }
