+++ src/main/java/janggi/domain/piece/StepPatternPiece.java
@@ -4,7 +4,7 @@ import janggi.domain.BoardInterface;
 import janggi.domain.Movement;
 import janggi.domain.Position;
 import janggi.domain.Side;
-import janggi.domain.strategy.MoveStrategy;
+import janggi.domain.policy.RoutePolicy;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,8 +12,8 @@ import java.util.List;
 public abstract class StepPatternPiece extends BasePiece {
     private final List<List<Movement>> moveRange;
 
-    public StepPatternPiece(List<List<Movement>> moveRange, MoveStrategy moveStrategy, Side side) {
-        super(moveStrategy, side);
+    public StepPatternPiece(List<List<Movement>> moveRange, RoutePolicy routePolicy, Side side) {
+        super(routePolicy, side);
         this.moveRange = moveRange;
     }
 
@@ -30,7 +30,7 @@ public abstract class StepPatternPiece extends BasePiece {
 
     @Override
     public boolean isMovable(List<Position> path, BoardInterface boardInterface) {
-        return moveStrategy.isMovable(path, side, boardInterface);
+        return routePolicy.isMovable(path, side, boardInterface);
     }
 
     private List<Position> calculatePath(Position start, List<Movement> path) {
