+++ src/main/java/janggi/domain/policy/JumpPolicy.java
@@ -1,4 +1,4 @@
-package janggi.domain.strategy;
+package janggi.domain.policy;
 
 import janggi.domain.BoardInterface;
 import janggi.domain.Position;
@@ -6,7 +6,7 @@ import janggi.domain.Side;
 
 import java.util.List;
 
-public class PoMoveStrategy implements MoveStrategy {
+public class JumpPolicy implements RoutePolicy {
     @Override
     public boolean isMovable(List<Position> path, Side side, BoardInterface boardInterface) {
         List<Position> piecesOnPath = path.stream()
