+++ src/main/java/janggi/domain/policy/RoutePolicy.java
@@ -1,4 +1,4 @@
-package janggi.domain.strategy;
+package janggi.domain.policy;
 
 import janggi.domain.BoardInterface;
 import janggi.domain.Position;
@@ -6,6 +6,6 @@ import janggi.domain.Side;
 
 import java.util.List;
 
-public interface MoveStrategy {
+public interface RoutePolicy {
     boolean isMovable(List<Position> path, Side side, BoardInterface boardInterface);
 }
